EXT_shader_framebuffer_fetch: Specify undefined behavior of helper invocations.
This seems like the most likely intention of the original extension
text, however because GLES 2.0 didn't explicitly define the concept of
helper invocation, the original text couldn't do better than
describing the issue in terms of the derivative functions dFdx, dFdy
and fwidth, which misses a number of cases (e.g. texturing functions)
that can potentially cause the value of a framebuffer fetch operation
performed by a helper invocation to have an influence on the result of
the fragment shader.

Re-specify the current undefined behavior in terms of helper
invocations where they are available, which should be more
future-proof than enumerating all operations that lead to cross-lane
communication in this extension, and has the additional advantage that
it doesn't forbid the use of such operations on a framebuffer fetch
result where the application is able to determine that there are no
helper invocations involved in the computation (e.g. via
gl_HelperInvocation).