EXT_texture_sRGB_decode: add interaction with ARB_bindless_texture
Per the discussion on API issue 51. Attempting to put the rules into
words became a mess, so I replaced it all with a table.

Note I also added the 'statically' which I believe came up in
earlier discussions on texelFetch, to make it clear that the
implementation is not required to solve the halting problem.