Define tokens for VIEW_CLASS_{EAC*,ETC2*,ASTC*}
Add view compatibility classes for EAC/ETC2 and ASTC formats
and add them for ARB_internalformat_query2.
These are mentioned in EXT/OES_texture_view for ES, but they
are never externally visibile and thus weren't assigned tokens.
However we also support these formats natively in GL on some
platforms, so we need tokens to return when the VIEW_COMPATIBILITY_CLASS
is queried on those platforms.

Reserve tokens for the VIEW_CLASSes for ASTC_3D, but don't
actually enable them since there are no extensions which
add these for GL.