Fix bogus use of enum group "TypeEnum"
I haven't done the archaeology to figure out what happened here. My
guess is that TypeEnum was intended as a catch-all for queriable
enumerant tokens, which are not realistically possible to list, and then
at some point it got used by the BeginConditionalRender commands.

While it isn't a complete fix, I changed the name of this group to
"ConditionalRenderMode" and used it only with those commands, and
removed the tagged of "TypeEnum" for the generic query commands.
Hopefully that won't subvert any outside use of the "TypeEnum" group.