Replaces #287, since I can't update @rbiermann's branch in-place in their repository.
@dgkoch @pdaniell-nv I suggest adding the KHR extension to this branch.

Unfortunately, the changes in #287 directly updated the headers rather than adding the new interfaces to gl.xml / wgl.xml. As a result regenerating the headers will not actually include the interfaces for the new extensions. If you can attempt to add the interfaces for the new extensions introduced in #287 that would be great. If not I'll attempt to do it, but I can't guarantee I'll have time by Sunday evening, with the Vulkan update taking up my time. In terms of just publishing the specs, the branch should be fine as is.