NV_pixel_buffer_object: fix example code
GL_STREAM_READ is not available in ES2.0. In fact, EXT_map_buffer_range
was the first time a mapped buffer could be read from legally.
Just use GL_STREAM_DRAW instead. This  is not a major problem, though,
as the usage hints do not affect the legal usage of a buffer and are
taken by the drivers merely as hints, sometimes even being ignored.

https://github.com/KhronosGroup/OpenGL-API/issues/66