Remove 'u' from NV_shader_atomic_uint64 GLASM option
Rename the GLASM option NV_shader_atomic_uint64 to
NV_shader_atomic_int64 (remove the 'u' in 'uint64').

The implementation of this extension checks for NV_shader_atomic_int64
and signals uint64 as invalid.

The extension itself is named without the 'u', and applications relying
on this would break if the driver is changed, so it's been opted to use
rename the option in this specification.